Add files via upload
Convert.py reads an excel sheet and makes the proper csv files for both BC and TC tiles. These csv files are separated by batches. To run change the name of "tiles.xlsx" to the .xlsx file with your tile data and use "python3 Convert.py". All csv files will be created and stored in the same directory as the script.